Rename all CSS classes that had UpperCamelCase CSS class names (for those which represented JavaScript classes), to lowercase.
This was done to remove the possibility of the problem that something is not a class name at some point (like "panel-header"), and then is upgraded to be its own class at a later date. To maintain consistency, the CSS class name would have to be renamed to "panel-Header", which cannot be done while maintaining backward compatibility.